Separate UpdateAllInstalledModsModal to its own component
